NULL: small error fixes.
The creation of the real hdf5 file didn't work properly because of a small error. Corrected it.

+++ Data/base_hdf5_creator.py
@@ -22,7 +22,7 @@ class VolumeInfo:
     """
 
     # VALUEs in km
-    dx: int
+    dx: float
     xt_min: float
     yt_min: float
     zt_min: float
